Add a way to test the gp2x renderer with SDL 2.0.
SDL 2.0's window pixel format is now 'take it or leave it', you
can't specify a the required format. Because the GP2X renderer
renders in 8bpp, then upmixes when scaling to 16-bit (the
format of choice on GP2X, presumably), we need to guarantee
16-bit output.

Add such support by allocating a shadow 16-bit framebuffer. The
renderer upmixes into this shadow, then the shadow is blitted
by SDL into the real window. This only happens if the real window
isn't already 16-bit.