Update nds, psp and wii ports.
Build fixes for latest (and by the looks of it, last) toolchains
released by the devkitpro project. Updates the NDS/Wii ports to
use the standard dirent API and drops the old custom libfat
implementation.

A bonus of this change is that the NDS build is working again
after being broken for more than two years. Yay!