Added per-user configuration file support
We may want to implement this in a better fashion, especially for
standalone releases. On Windows there is no problem, but on other
operating systems if each standalone game writes over the same
configuration file we will have some problems, particularly if their
include their own settings. The best approach may be to put this setting
in the global config file- read it first, then decide whether or not to
create a per-user config file.