Improved intake event processing.
Intake should now miss far fewer keystrokes due to waiting for the next
event and only consuming one event at a time, rather than using the
existing event infrastructure that will consume all events as they
appear. This required changing the wait_event function, which involved changing
things in the Wii and NDS ports. If these ports are worked on in the
future these may need to be testing.