Added msys2 build scripts.
These scripts are located in scripts/msys2-buildscripts and are designed
to allow for Windows builds to be created and uploaded to update servers
from a blank msys2 installation. It is highly recommended that you create
a new msys2 installation and only run these scripts from there. See
BUILD SCRIPTS.txt for more information.