Split SUPPRESS_BUILD into several flags, use with "make dist" (#268)
Replaces `SUPPRESS_BUILD` with new flags to selectively disable parts of the default Makefile targets. This allows the platforms that use a `dist` target to disable various built-in MZX Makefile targets to avoid bugs and/or confusion.

* `SUPPRESS_ALL_TARGETS`: does the same thing `SUPPRESS_BUILD` previously did. Only useful for the `none` platform.
* `SUPPRESS_CC_TARGETS`: prevents adding `mzx`, `utils`, etc. to `all`. Useful for platforms that have a `dist` rule so it's harder to accidentally start a single-ABI build when the general use case would be to use a meta target to generate builds for multiple ABIs.
* `SUPPRESS_BUILD_TARGETS`: prevents `build` and its related targets from being defined. Only really useful for Android, where the usual build process just doesn't work at all.
* `SUPPRESS_HOST_TARGETS`: prevents `test`, `assets/help.fil`, etc. targets from being defined. These targets rely on being able to run target executables (either `mzxrun` or one of the utils) on the current host, which is really only applicable for the `mingw`, `unix*`, `darwin*`, `amiga`, `pandora` platforms.

Additionally, the pefix rules now clearly indicate that they use HOST_CC instead of the regular CC and don't append ".exe" to the executable target (which is confusing when cross-compiling MinGW builds).

Additionally, the "recently redesigned" comment in the Makefile that's been there since 2007 has been tweaked.