Add joystick config unit test, fix joystick ranges for NDS/3DS (#269)
Fixes a regression caused by e42259b4 where joystick config ranges
(example: joy[1,16].a = key_space) stopped working for the NDS and 3DS
ports. This is a problem as game configs should be using ranges like
these for portability.

The bounding check for the joystick mapping functions is now entirely
based on the first joystick index (the last index is just clamped to
the range of valid joysticks). This means any joystick range starting
from 1 should be guaranteed to work on everything. A unit test for the
joystick configuration options has also been added.