Remove invalid copyrights from various files.
An inspection of the following files and their equivalents prior to
them being split from clipboard.c shows that no substantial
modifications were made to any of their contents. The Win32 and X11
clipboard handlers are line-for-line identical to their implementations
prior to Nov 28th 2017; clipboard.h is line-for-line identical to
how it existed when I created it; and the NULL clipboard handler
received trivial modifications that have been reverted.

* clipboard.h
* clipboard_null.h
* clipboard_win32.h
* clipboard_x11.h

Compare to files prior to commit f2040a6:

* https://github.com/AliceLR/megazeux/blob/ce519eccb7ca828d9f7761fefdf4aa31a464308b/src/editor/clipboard.h
* https://github.com/AliceLR/megazeux/blob/ce519eccb7ca828d9f7761fefdf4aa31a464308b/src/editor/clipboard.c