Add renderer hook for DJGPP hardware SMZX mode. (#384)
The DJGPP port EGA renderer handled hardware SMZX mode in its
`set_video_mode` function, which only gets called when initializing
a renderer or creating a new window (as it corresponds to the SDL 1.x
function of the same name). This patch adds an extra hook specifically
for `set_screen_mode` to be called in `update_screen` prior to
committing palette changes to the renderer.