Add compatibility for MZX 1.x reverse send, add 1.x/2.x scan tests. (#438)
MegaZeux 1.x lacks the `robot::status` field and instead stores
label sent flags in a global array, which robots set in *any* robot
that receives a label (unconditionally) and clear at the start of
execution and during some commands (presumably `GOTO`/`SEND`/`IF`
so they don't run again). The reverse scan will start execution for
any robot with this flag set, similar to when `status` == 2 in
MegaZeux 2.x. This patch hacks support for this behavior onto the
existing `robot::status` field.