Add Robotic *_prefix constants, fix REL COUNTERS/COPY behavior for 1.x (#442)
The combination of `REL COUNTERS` and `COPY` behaved sensibly in
MegaZeux 1.x—it applied `XPOS` and `YPOS` to both sets of coordinates,
which is the same as what the 2.00 help file *claimed* to do. However,
this is not how MegaZeux 2.x actually behaves, so the old behavior
needed a compatibility hack.