Fix RAD nested riff precedence bug.
If a playing channel riff invokes a riff, the new riff will replace
the current playing riff or stop the current playing riff (if it is
R0x/T0x or Rx0/Tx0). This happens unless the riff recursion limit is
reached, in which case the Rxx/Txx is ignored entirely. TickRiff
needs to exit early when this occurs to prevent it from corrupting
the new riff.