Mark all new worlds created in the editor as active.
The editor would mark the blank world created on initialization as
active, but not any blank worlds created afterward during editing.
This would cause blank worlds created by Alt+R and failed __TEST.MZX
reloads to be "inactive", which as of 2.91g, has special handling
in the game.c draw function to skip drawing the screen and instead
constantly display the intro timer.

The fix is to just set this flag in create_blank_world where it
should have been doing it anyway.