Update Emscripten to 4.0.7 and switch default SDL to SDL3. (#539)
No issues encountered after a reasonable amount of testing in Firefox;
going to switch the default for now and revert if issues are found later.
This exposed a preexisting bug where emzip was relying on SDL
functions. Switching it to the utils vio.c and utils_alloc.h fixed this.