Fix slow softscale rendering with direct3d11 driver. (#541)
The adoption of faster software rendering algorithms means that
MegaZeux's software renderer is reading from the memory it is
writing to more often. softscale was rendering to the pointer
returned by `SDL_LockTexture`, which SDL generally doesn't want to
be read (as modern APIs use mapping that is slow to read from).

+ `direct3d`, `software`, PS2, PSP, Vita, and Wii U render drivers
  will continue to use `SDL_LockTexture` for now, as it is significantly
  faster on old systems and consoles.
+ All other drivers now instead use `SDL_UpdateTexture`, which is
  identical in performance for opengl drivers, negligible for Metal(?),
  faster for ogc (Wii/GC), and avoids extremely slow performance
  caused by write-only mapping with modern API drivers.
+ The default render driver for Windows is now explicitly `direct3d`
  as it is faster for old machines than the SDL3 default `direct3d11`.

SDL3 support is now marked as OK for mingw and enabled for release builds.